changes requested

+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/production/turbines/GT_MTE_LargeTurbine_SHSteam.java
@@ -80,7 +80,7 @@ public class GT_MTE_LargeTurbine_SHSteam extends GregtechMetaTileEntity_LargerTu
         int remainingFlow = MathUtils.safeInt((long) (aOptFlow
                 * 1.25f)); // Allowed to use up to 125% of optimal flow.  Variable required outside of loop for
         // multi-hatch scenarios.
-        this.realOptFlow = aOptFlow * flowMultipliers[0];
+        this.realOptFlow = (double) aOptFlow * (double) flowMultipliers[0];
 
         storedFluid = 0;
         for (int i = 0; i < aFluids.size() && remainingFlow > 0; i++) {
